+++ src/main/java/plugily/projects/buildbattle/handlers/reward/Reward.java
@@ -102,7 +102,7 @@ public class Reward {
 
   public enum RewardType {
     END_GAME("endgame"), GTB_GUESS("guess"), GTB_ALL_GUESSED("allguessed"), VOTE("vote"), REPORT("report"), PLACE("place"),
-    START_GAME("startgame"), SCOREBOARD_REMOVED("scoreboard_remove"), SCOREBOARD_ADD("scoreboard_add");
+    START_GAME("startgame");
 
     private final String path;